Extract gateway QoS catalog helpers
